Move encoding utils to encoding.py
